Support unpacking from streams.
This is perhaps slightly janky, but hopefully quite flexible:

Each `DataType` now has a `stream_unpack(stream)` method that returns
`True` or `False` to indicate whether the value has been completely
unpacked.

`stream` is an instance of `StreamUnpacker`. When `value` is not
completely unpacked, it may optinally save a state to `stream` via
`stream.push_state(self, state)`. On subsequent calls to
`unpack_stream(stream)` with the same `stream` object passed,
`unpack_stream()` will recall the saved state from the stack and
continue unpacking.

`StreamUnpacker` is simply a buffer and a generator. It is instantiated
with a type, e.g. `stream = StreamUnpacker(some_struct)` and then
`stream.unpack(buf)` is a generator yielding `some_struct` instances
from the pased buffer.

On each call to `stream.unpack(buf)`, 0 or more instances may be
yielded. The passed buffer is concatenated to the stream's internal
buffer, and any unpacked bytes remain in the internal buffer between
calls. Example:

    stream = StreamUnpacker(my_struct)
    sock = socket.socket(socket.ATF_INET, socket.SOCK_STREAM)
    sock.connect(...)

    while True:
	for obj in stream.unpack(sock.recv(4096)):
	    print('received', obj)